+++ ordinal-map/src/ordinal/impls/mod.rs
@@ -2,6 +2,7 @@ mod array;
 mod boxed;
 mod either;
 mod non_zero;
+mod ordering;
 mod primitive;
 mod std;
 mod tuple;